refactor : blackgame 객체를 통한 응집도 높이기
