+++ src/main/kotlin/blackjack/BlackjackApplication.kt
@@ -1,10 +1,7 @@
 package blackjack
 
-import blackjack.dto.PlayersDto
 import blackjack.dto.ResultDto
 
-const val TOP_CARD = 0
-const val BASIC_RULE_COUNT = 2
 const val BASIC_RULE_DELIMITER = ","
 
 val cardShuffle = BlackjackCardShuffle()
@@ -13,7 +10,6 @@ val cards = cardShuffle.getCards().toMutableList()
 fun main() {
     val gamers = mutableListOf<UserRole>(Dealer())
 
-
     println("게임에 참여할 사람의 이름을 입력하세요.(쉼표 기준으로 분리)")
     val players = readln()
         .split(BASIC_RULE_DELIMITER)
@@ -21,42 +17,9 @@ fun main() {
 
     gamers.addAll(players)
 
-    repeat(BASIC_RULE_COUNT) {
-        gamers.forEach {
-            it.draw(cards.removeAt(TOP_CARD))
-        }
-    }
+    val blackjackGame = BlackjackGame(gamers)
 
-    val playersDto = PlayersDto(gamers)
-    playersDto.showPlayerNames()
-    playersDto.showInitCards()
-
-    val results = mutableListOf<UserRole>()
-    for (player in players) {
-        var p = player
-        while (!p.isFinish()) {
-            p = deal(p) as Player
-        }
-        results.add(p)
-    }
-
-    var dealer = gamers.first { it.isDealer() }
-
-
-    while (!dealer.isFinish()) {
-        dealer = dealer.draw(cards.removeAt(0))
-    }
-
-    results.add(dealer)
-
-    val resultDto = ResultDto(results)
+    val resultDto = ResultDto(blackjackGame.play())
     resultDto.result()
 }
 
-private fun deal(player: Player): UserRole {
-    println("%s 님은 한장의 카드를 더 받겠습니까?(예는 y, 아니오는 n)".format(player.name))
-    return when (readln()) {
-        "y" -> player.draw(cards.removeAt(0))
-        else -> player.stand()
-    }
-}
