+++ src/main/kotlin/blackjack/BlackjackGame.kt
@@ -1,0 +1,58 @@
+package blackjack
+
+import blackjack.dto.PlayersDto
+
+class BlackjackGame(private val gamers: List<UserRole>) {
+
+    init {
+        repeat(BASIC_RULE_COUNT) {
+            gamers.forEach {
+                it.draw(cards.removeAt(TOP_CARD))
+            }
+        }
+
+        val playersDto = PlayersDto(gamers)
+        playersDto.showPlayerNames()
+        playersDto.showInitCards()
+    }
+
+    fun play(): List<UserRole> {
+        val players = gamers.filter { !it.isDealer() }
+        val results = mutableListOf<UserRole>()
+        for (player in players) {
+            results.add(playPlayerTurn(player))
+        }
+
+        results.add(playDealerTurn())
+        return results.toList()
+    }
+
+    private fun playPlayerTurn(player: UserRole): UserRole {
+        var p = player
+        while (!p.isFinish()) {
+            p = deal(p) as Player
+        }
+        return p
+    }
+
+    private fun playDealerTurn(): UserRole {
+        var dealer = gamers.first { it.isDealer() }
+        while (!dealer.isFinish()) {
+            dealer = dealer.draw(cards.removeAt(TOP_CARD))
+        }
+        return dealer
+    }
+
+    private fun deal(player: UserRole): UserRole {
+        println("%s 님은 한장의 카드를 더 받겠습니까?(예는 y, 아니오는 n)".format(player.name))
+        return when (readln()) {
+            "y" -> player.draw(cards.removeAt(TOP_CARD))
+            else -> player.stand()
+        }
+    }
+
+    companion object {
+        private const val BASIC_RULE_COUNT = 2
+        const val TOP_CARD = 0
+    }
+}
